Rename GeneralRelativity Tags
